enrollment page student

+++ front-end/src/app/components/delete/delete.component.ts
@@ -3,6 +3,7 @@ import { Component, OnInit } from '@angular/core';
 import { MdbModalRef } from 'mdb-angular-ui-kit/modal';
 import { UserService } from 'src/app/services/user.service';
 import { GlobalStudent } from 'src/app/services/Global.student.service';
+import { ConfirmationService } from 'src/app/services/confirmation.service';
 
 @Component({
   selector: 'app-delete',
@@ -10,7 +11,11 @@ import { GlobalStudent } from 'src/app/services/Global.student.service';
   styleUrls: ['./delete.component.css'],
 })
 export class DeleteComponent implements OnInit {
+  action: string | null = null;
+  body: string | null = null;
+
   constructor(
+    private confirmPublish: ConfirmationService,
     public modalRef: MdbModalRef<DeleteComponent>,
     public GlobalStudent: GlobalStudent,
     private userService: UserService,
@@ -20,16 +25,21 @@ export class DeleteComponent implements OnInit {
   ngOnInit(): void {}
 
   delete = () => {
-    this.userService.deleteAcc().subscribe((response: any) => {
-      if (response.error) {
-        console.log('error while deleting the account');
-        return;
-      }
+    if (this.action == 'account') {
+      this.userService.deleteAcc().subscribe((response: any) => {
+        if (response.error) {
+          console.log('error while deleting the account');
+          return;
+        }
 
-      this.removeStudent();
+        this.removeStudent();
+        this.modalRef.close();
+        this.router.navigate(['/home']);
+      });
+    } else if (this.action == 'enrollment') {
+      this.confirmPublish.updateGlobalVarDE(true);
       this.modalRef.close();
-      this.router.navigate(['/home']);
-    });
+    }
   };
 
   removeStudent() {
